Improved error messages and modularity

+++ src/main.rs
@@ -1,27 +1,31 @@
 use std::fs;
 use std::path::Path;
 use std::ffi::OsStr;
-use std::ffi::OsString;
+use std::error::Error;
 
 fn main() {
 
     let mypath = "C:/Users/dhjay/Music/Music/";
 
+    // Set flle extensions to keep
     let file_extensions = vec![OsStr::new("flac"), OsStr::new("mp3"), OsStr::new("webm")];
+
+    println!("Found:");
+
     // Scan files and folders in directory
     let (files, folders) = match scan_path(mypath) {
 
         Ok((fi, fo)) => {
         
-            println!("\nFound {} files:\n", fi.len());
-            for f in &fi {
-                println!("{:?}", f.file_name());
-            }
+            println!("  => {} files", fi.len());
+            // for f in &fi {
+            //     println!("{:?}", f.file_name());
+            // }
 
-            println!("\nFound {} folders:\n", fo.len());
-            for f in &fo {
-                println!("{:?}", f.file_name());
-            }
+            println!("  => {} folders", fo.len());
+            // for f in &fo {
+            //     println!("{:?}", f.file_name());
+            // }
 
             (fi, fo)
         },
@@ -30,38 +34,28 @@ fn main() {
 
     };
 
-    // Recursively scan folders for flacs
+    // Recursively scan folders for flacs (UNSAFE)
     let mut deep_files: Vec<fs::DirEntry> = Vec::new();
     recursive_find(&folders, &mut deep_files);
-    println!("\nFound {} deep files:\n", deep_files.len());
-    for f in &deep_files {
-        println!("{:?}", f.file_name());
-    }
+    println!("  => {} files nested in folders", deep_files.len());
+    // for f in &deep_files {
+    //     println!("{:?}", f.file_name());
+    // }
 
-    println!("\nOf which are music:\n");
-    for file in deep_files {
-        let path = file.path();
-        let path = Path::new(&path);
-        match path.extension() {
-            Some(val) => {
-                if file_extensions.contains(&val) {
-                    println!("{}", file.file_name().into_string().unwrap());
-                    let destination = format!("{}{}", mypath, file.file_name().into_string().unwrap());
-
-                    // Copy over
-                    fs::copy(file.path(), destination).unwrap();
-                }
-            },
-            None => ()
-        };
+    println!("\nExtracting...");
+    if let Err(e) = extract_music(&deep_files, &file_extensions, String::from(mypath)) {
+        panic!("ERROR: {}", e);
     }
 
     // Remove folders
+    println!("Removing folders...");
     for folder in folders {
         if !folder.file_name().into_string().unwrap().starts_with(".") {
             fs::remove_dir_all(folder.path()).unwrap();
         }
     }
+
+    println!("Complete!");
 }
 
 fn scan_path(directory: &str) -> Result<(Vec<fs::DirEntry>, Vec<fs::DirEntry>), std::io::Error> {
@@ -93,6 +87,7 @@ fn recursive_find(folders: &Vec<fs::DirEntry>, found_files: &mut Vec<fs::DirEntr
         return
     }
     for folder in folders {
+        
         let (deep_files, deep_folders) = scan_path(folder.path().to_str().unwrap()).unwrap();
         for file in deep_files {
             found_files.push(file);
@@ -101,4 +96,25 @@ fn recursive_find(folders: &Vec<fs::DirEntry>, found_files: &mut Vec<fs::DirEntr
         recursive_find(&deep_folders, found_files);
     }
 
+}
+
+fn extract_music(files: &Vec<fs::DirEntry>, file_extensions: &Vec<&OsStr>, origin: String) -> Result<(), Box<dyn Error>> {
+
+    for file in files {
+        let path = file.path();
+        let path = Path::new(&path);
+        match path.extension() {
+            Some(val) => {
+                if file_extensions.contains(&val) {
+                    // println!("{}", file.file_name().into_string().unwrap());
+                    let destination = format!("{}{}", origin, file.file_name().into_string().unwrap());
+                    // Copy over
+                    fs::copy(file.path(), destination)?;
+                }
+            },
+            None => ()
+        };
+    }
+
+    Ok(())
 }
